src/test/javascript/spec/app/entities/offer-manage/offer-manage-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { ELearningTestModule } from '../../../test.module';
import { OfferManageUpdateComponent } from 'app/entities/offer-manage/offer-manage-update.component';
import { OfferManageService } from 'app/entities/offer-manage/offer-manage.service';
import { OfferManage } from 'app/shared/model/offer-manage.model';

describe('Component Tests', () => {
    describe('OfferManage Management Update Component', () => {
        let comp: OfferManageUpdateComponent;
        let fixture: ComponentFixture<OfferManageUpdateComponent>;
        let service: OfferManageService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [ELearningTestModule],
                declarations: [OfferManageUpdateComponent]
            })
                .overrideTemplate(OfferManageUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(OfferManageUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(OfferManageService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new OfferManage(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.offer = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new OfferManage();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.offer = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
